requirements.txt
tk
psutil
datetime
requests
pillow
python-dotenv